Se implementaron algunos cambios en la forma de funcionar del parpadeo, tal como se explica en el registro de testeos, pero no parece producir alguna mejoría significativa. Versión: 1.0.14.

+++ main/Core/Src/main.c
@@ -583,7 +583,7 @@ void debounce_task(void *argument)
 {
   /* USER CODE BEGIN debounce_task */
 
-  const uint8_t debounce_time = 200; //Tiempo prudente de espera por rebotes.
+  const int debounce_time = 200; //Tiempo prudente de espera por rebotes.
 
   /* Infinite loop */
   for(;;)
